+++ packages/rewild-renderer/lib/renderers/sky/CloudsRenderer.ts
@@ -133,7 +133,7 @@ export class CloudsRenderer {
     cloudPass.setPipeline(this.pipeline);
     cloudPass.setVertexBuffer(0, geometry!.vertexBuffer);
     cloudPass.setVertexBuffer(1, geometry!.uvBuffer);
-    cloudPass.setIndexBuffer(geometry!.indexBuffer, 'uint16');
+    cloudPass.setIndexBuffer(geometry!.indexBuffer, 'uint32');
     cloudPass.setBindGroup(0, this.bindGroup);
     cloudPass.drawIndexed(geometry!.indices!.length);
     cloudPass.end();
